refactor(@dpc-sdp/ripple-tide-event): :art: migrate tide-event to use nuxt layers
