+++ package.json
@@ -17,7 +17,7 @@
     "dev": "yarn run build && concurrently \"yarn storybook\" \"yarn workspace nuxt-app dev\"",
     "clean": "git clean -dfx && git reset --hard && yarn",
     "build": "lerna run build --ignore docs",
-    "build:ripple": "lerna run build --scope @dpc-sdp/ripple*",
+    "build:ripple": "lerna run build --scope @dpc-sdp/ripple-ui-* --scope @dpc-sdp/ripple-tide-api",
     "build:examples": "lerna run build --scope nuxt-app --scope vue-app --scope webcomponents",
     "build:docs": "lerna run build --scope docs",
     "build:storybook": "build-storybook",
