+++ packages/nuxt-ripple/package.json
@@ -1,31 +1,12 @@
 {
-  "private": true,
   "license": "Apache-2.0",
   "name": "@dpc-sdp/nuxt-ripple",
-  "version": "2.1.0-alpha.31",
-  "scripts": {
-    "dev": "nuxi dev",
-    "dev:debug": "node --inspect node_modules/.bin/nuxi dev",
-    "build": "nuxi build",
-    "compile": "tsc --build",
-    "start": "node .output/server/index.mjs",
-    "mockserver": "node ./mockserver.js",
-    "start:mock": "API_URL=http://localhost:3001 concurrently 'yarn mockserver' 'yarn start'",
-    "cy:open": "cypress open --e2e --browser electron",
-    "cy:run": "cypress run",
-    "test:integration": "start-test 'NUXT_PUBLIC_API_URL=http://localhost:3001 yarn dev' http-get://localhost:3000/api/tide/site?id=8888 'cy:open'"
-  },
+  "main": "./nuxt.config.ts",
+  "version": "2.1.0-alpha.52",
   "dependencies": {
-    "@dpc-sdp/ripple-tide-api": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-tide-event": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-tide-grant": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-tide-landing-page": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-tide-media": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-tide-news": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-tide-publication": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-tide-publication-page": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-ui-core": "2.1.0-alpha.31",
-    "@dpc-sdp/ripple-ui-forms": "2.1.0-alpha.31"
+    "@dpc-sdp/ripple-tide-api": "2.1.0-alpha.52",
+    "@dpc-sdp/ripple-ui-core": "2.1.0-alpha.52",
+    "@dpc-sdp/ripple-ui-forms": "2.1.0-alpha.52"
   },
   "peerDependencies": {
     "nuxt": "3.0.0"
