+++ packages/ripple-tide-event/components/global/TideEvent.vue
@@ -1,16 +1,6 @@
-<script lang="ts">
-export default { name: 'TideEventPage' }
-</script>
-
 <template>
-  <TideBaseLayout
-    :site="site"
-    :background="page.background"
-    :pageTitle="page.title"
-    :pageDescription="page.description"
-    :pageLanguage="page.lang"
-    :updatedDate="page.changed || page.created"
-  >
+  <TideBaseLayout :site="site" :background="page.background" :pageTitle="page.title" :pageDescription="page.description"
+    :pageLanguage="page.lang" :updatedDate="page.changed || page.created">
     <template #aboveHeader>
       <slot name="aboveHeader"></slot>
     </template>
@@ -21,20 +11,11 @@ export default { name: 'TideEventPage' }
       <slot name="breadcrumbs"></slot>
     </template>
     <template #aboveBody="{ hasBreadcrumbs }">
-      <TideEventHeader
-        :header="page.header"
-        :hasBreadcrumbs="hasBreadcrumbs"
-      ></TideEventHeader>
+      <TideEventHeader :header="page.header" :hasBreadcrumbs="hasBreadcrumbs"></TideEventHeader>
     </template>
     <template #body>
-      <TideEventOverview
-        :date="page.date"
-        :showTime="page.showTime"
-        :overview="page.overview"
-        :details="page.details"
-        :link="page.link"
-        :description="page.description"
-      ></TideEventOverview>
+      <TideEventOverview :date="page.date" :showTime="page.showTime" :overview="page.overview" :details="page.details"
+        :link="page.link" :description="page.description"></TideEventOverview>
       <TideEventBody :body="page.body" :link="page.link"></TideEventBody>
     </template>
     <template #aboveSidebar>
@@ -54,7 +35,7 @@ export default { name: 'TideEventPage' }
 
 <script setup lang="ts">
 import { TideSiteData } from '@dpc-sdp/ripple-tide-api/types'
-import type { TideEventPage } from './types'
+import type { TideEventPage } from './../../types'
 
 interface Props {
   site: TideSiteData
