+++ packages/ripple-tide-event/components/tide-event-body.vue
