+++ packages/ripple-tide-event/package.json
@@ -1,33 +1,13 @@
 {
+  "license": "Apache-2.0",
   "name": "@dpc-sdp/ripple-tide-event",
-  "version": "2.1.0-alpha.52",
   "description": "Ripple mappings and components for Tide Event Content type",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
-  "type": "module",
-  "exports": {
-    ".": "./dist/index.js",
-    "./component": "./src/index.vue",
-    "./types": "./src/types.ts",
-    "./nuxt": "./src/nuxt.ts"
-  },
-  "license": "MIT",
-  "keywords": [],
-  "scripts": {
-    "clean": "rimraf dist* && rimraf tsconfig.tsbuildinfo",
-    "build": "yarn clean && tsc -p tsconfig.json",
-    "watch": "yarn clean && tsc -p tsconfig.json -w"
-  },
+  "main": "./nuxt.config.ts",
+  "version": "2.1.0-alpha.52",
   "dependencies": {
-    "@dpc-sdp/ripple-tide-api": "2.1.0-alpha.52"
+    "@dpc-sdp/nuxt-ripple": "2.1.0-alpha.52"
   },
-  "files": [
-    "dist",
-    "src",
-    "!**/*.test.*",
-    "!**/*.json",
-    "CHANGELOG.md",
-    "LICENSE",
-    "README.md"
-  ]
+  "peerDependencies": {
+    "nuxt": "3.0.0"
+  }
 }
