+++ packages/ripple-tide-event/server/middleware/event.ts
@@ -1,0 +1,7 @@
+import { defineEventHandler } from 'h3'
+import { registerTideContentType } from '@dpc-sdp/ripple-tide-api'
+import tideEventModule from '../../mapping'
+
+export default defineEventHandler(async (event) => {
+  registerTideContentType(event, 'event', tideEventModule)
+})
