--- packages/ripple-tide-event/src/nuxt.ts
@@ -1,26 +1,0 @@
-import { join } from 'pathe'
-import { defineNuxtModule, addComponentsDir, addComponent } from '@nuxt/kit'
-
-export default defineNuxtModule({
-  async setup(_opt, nuxt) {
-    // Add TideEvent component and mapping
-    addComponent({
-      name: 'TideEventPage',
-      filePath: join(__dirname, './index.vue'),
-      global: true
-    })
-    nuxt.options.runtimeConfig.public.tide.mapping.content.event = join(
-      __dirname,
-      '../dist/index.js'
-    )
-
-    // Add TideEvent page components as dynamic imports for Nuxt apps - See https://v3.nuxtjs.org/guide/concepts/auto-imports
-    addComponentsDir({
-      extensions: ['vue'],
-      path: join(__dirname, './components'),
-      prefix: 'TideEvent',
-      global: true
-    })
-    console.log('Added TideEvent module')
-  }
-})
