--- packages/ripple-tide-event/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "noImplicitAny": false,
-    "declaration": false,
-    "jsx": "preserve",
-    "paths": {
-      "@dpc-sdp/ripple-tide-api": ["./../ripple-tide-api/dist/index.js"]
-    }
-  },
-  "include": ["./src"]
-}
